# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "cf-teamspeed API",
    "version": "1.0.0",
    "description": "基于 Cloudflare Realtime SFU + Workers 的私人游戏语音室 API。\n\n**这份文档覆盖完整的核心闭环**（客户端最少需要实现的部分）：配置、鉴权与续期、房间生命周期、WebSocket 握手票、SFU 信令、在线状态。\n\n其余接口（房间管理、管理后台）见 `docs/API.md`，那份是完整清单；本文件刻意保持精简，以便作为「新客户端接入检查表」使用。\n\n鉴权：除 `security: []` 的接口外，一律带 `Authorization: Bearer <token>`。\n\n错误：统一返回 `{ok:false, error, code}`；`code` 是稳定契约，分支判断请用它而不是 `error` 文案。",
    "license": { "name": "MIT" }
  },
  "servers": [
    { "url": "https://ts.futurvo.cc", "description": "示例部署（请替换成你自己的 Worker 域名）" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "登录接口返回的 `token`。有效期 12 小时，随请求滚动续期；服务端会通过响应头 `X-Refreshed-Token` 下发新 token。"
      }
    },
...
    "schemas": {
      "ApiOk": {
        "type": "object",
        "required": ["ok", "data"],
        "properties": {
          "ok": { "const": true },
          "data": {}
        }
      },
      "ApiError": {
        "type": "object",
        "required": ["ok", "error"],
        "properties": {
          "ok": { "const": false },
          "error": { "type": "string", "description": "给用户看的提示；文案可能变动" },
          "code": { "$ref": "#/components/schemas/ErrorCode" },
          "detail": { "description": "调试信息" }
        }
      },
      "ErrorCode": {
        "type": "string",
        "description": "稳定错误码。客户端分支逻辑必须用它。",
        "enum": [
          "INVALID_KEY",
          "INVALID_BODY",
          "NICKNAME_TAKEN",
          "NICKNAME_RESERVED",
          "NICKNAME_INVALID",
          "UNAUTHORIZED",
          "SESSION_EXPIRED",
          "FORBIDDEN",
          "ROOM_NOT_FOUND",
          "ROOM_FULL",
          "RATE_LIMITED",
          "BANNED",
          "TURNSTILE_FAILED",
          "BAD_TICKET",
          "INTERNAL"
        ]
      },
...
      "RoomSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "ownerUid": { "type": ["string", "null"] },
          "memberCount": { "type": "integer" },
          "maxMembers": { "type": "integer" },
          "createdAt": { "type": "integer" }
        }
      },
      "RoomMember": {
        "type": "object",
        "properties": {
          "uid": { "type": "string" },
          "nickname": { "type": "string" },
          "avatarId": { "type": ["string", "null"] },
          "avatarUrl": { "type": ["string", "null"] },
          "role": { "type": "string", "enum": ["guest", "admin"] },
          "muted": { "type": "boolean" },
          "joinedAt": { "type": "integer" },
          "lastSeen": { "type": "integer" }
        }
      },
      "RoomSnapshot": {
        "type": "object",
        "properties": {
          "room": { "$ref": "#/components/schemas/RoomSummary" },
          "members": { "type": "array", "items": { "$ref": "#/components/schemas/RoomMember" } },
          "version": { "type": "integer", "description": "每次变更自增" }
        }
      },
      "WsTicket": {
        "type": "object",
        "required": ["ticket", "expiresAt", "wsUrl"],
        "properties": {
          "ticket": { "type": "string", "description": "一次性票据，60 秒有效" },
          "expiresAt": { "type": "integer" },
          "wsUrl": { "type": "string", "description": "已拼好的 WebSocket 地址，直接用它连接" }
        }
      },
      "TrackRef": {
        "type": "object",
        "properties": {
          "trackName": { "type": "string" },
          "uid": { "type": "string", "description": "发布者" },
          "sessionId": { "type": "string", "description": "发布者 sessionId，订阅时要用" },
          "mid": { "type": ["string", "null"] }
        }
      },
...
    },
    "responses": {
      "Unauthorized": {
        "description": "未登录 / token 无效 / 会话到顶（`code` 分别为 `UNAUTHORIZED`、`SESSION_EXPIRED`）",
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
        }
      },
      "RateLimited": {
        "description": "请求过于频繁，按 `Retry-After` 头退避",
        "headers": { "Retry-After": { "schema": { "type": "integer" }, "description": "建议等待秒数" } },
        "content": {
          "application/json": { "schema": { "$ref": "#/components/schemas/ApiError" } }
        }
      }
    }
  },
  "paths": {
...
    "/api/rooms": {
      "get": {
        "tags": ["rooms"],
        "summary": "房间列表（含成员明细）",
        "responses": {
          "200": {
            "description": "房间列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "rooms": {
                              "type": "array",
                              "items": {
                                "allOf": [
                                  { "$ref": "#/components/schemas/RoomSummary" },
                                  {
                                    "type": "object",
                                    "properties": {
                                      "members": { "type": "array", "items": { "$ref": "#/components/schemas/RoomMember" } }
                                    }
                                  }
                                ]
                              }
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      },
      "post": {
        "tags": ["rooms"],
        "summary": "创建房间",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name"],
                "properties": {
                  "name": { "type": "string", "maxLength": 24 },
                  "maxMembers": { "type": "integer", "minimum": 2, "maximum": 50 }
                }
              }
            }
          }
        },
        "responses": { "201": { "description": "已创建" }, "429": { "$ref": "#/components/responses/RateLimited" } }
      }
    },
    "/api/rooms/default": {
      "get": {
        "tags": ["rooms"],
        "summary": "取默认房间 home（不存在则创建）",
        "responses": { "200": { "description": "默认房间" }, "401": { "$ref": "#/components/responses/Unauthorized" } }
      }
    },
    "/api/rooms/{id}/snapshot": {
      "get": {
        "tags": ["rooms"],
        "summary": "房间快照（权威状态）",
        "description": "收到 WebSocket 的 `room-changed` 后调它拿最新状态。",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "快照",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/RoomSnapshot" } } }
                  ]
                }
              }
            }
          },
          "404": { "description": "房间不存在（`ROOM_NOT_FOUND`）" }
        }
      }
    },
    "/api/rooms/{id}/tracks": {
      "get": {
        "tags": ["rooms"],
        "summary": "房间内可订阅的轨道",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "轨道列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "tracks": { "type": "array", "items": { "$ref": "#/components/schemas/TrackRef" } }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{id}/join": {
      "post": {
        "tags": ["rooms"],
        "summary": "进入房间",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "已进入",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "snapshot": { "$ref": "#/components/schemas/RoomSnapshot" },
                            "keyId": { "type": "string", "description": "房间密钥 id（E2EE 用）" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "409": { "description": "房间已满（`ROOM_FULL`）" },
          "429": { "$ref": "#/components/responses/RateLimited" }
        }
      }
    },
    "/api/rooms/{id}/leave": {
      "post": {
        "tags": ["rooms"],
        "summary": "离开房间",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "已离开" } }
      }
    },
    "/api/rooms/{id}/heartbeat": {
      "post": {
        "tags": ["rooms"],
        "summary": "HTTP 心跳",
        "description": "WebSocket 可用时优先走 socket 心跳。超过 45 秒无任何信号会被判定离线并回收房间资源。",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["roomId"],
                "properties": { "roomId": { "type": "string" } }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "心跳结果",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "version": { "type": "integer" },
                            "reaped": { "type": "integer", "description": "本次顺带清理的离线成员数" }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{id}/mute": {
      "post": {
        "tags": ["rooms"],
        "summary": "上报麦克风状态",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": { "muted": { "type": "boolean" } }
              }
            }
          }
        },
        "responses": { "200": { "description": "已更新" } }
      }
    },
    "/api/rooms/{id}/ws-ticket": {
      "post": {
        "tags": ["rooms"],
        "summary": "领 WebSocket 握手票",
        "description": "WebSocket API 无法自定义请求头，所以必须先走这个 HTTP 接口拿一次性票据，再用返回的 `wsUrl` 连接。票据 60 秒有效、单次有效，每次重连都要重新领。",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "票据",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/ApiOk" },
                    { "type": "object", "properties": { "data": { "$ref": "#/components/schemas/WsTicket" } } }
                  ]
                }
              }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" }
        }
      }
    },
...
  },
...
}